freezed_requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cn2an==0.5.23
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
crcmod==1.7
cryptography==44.0.2
ctranslate2==4.6.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Distance==0.1.3
distro==1.9.0
editdistance==0.8.1
einops==0.8.1
evdev==1.9.1
executing==2.2.0
fastapi==0.115.12
faster-whisper==1.1.1
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
funasr==1.2.6
future==1.0.0
g2p-en==2.1.0
gradio==5.32.0
gradio_client==1.10.2
groovy==0.1.2
h11==0.14.0
hf-xet==1.1.2
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
jieba_fast==0.53
Jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kaldiio==2.18.1
kiwisolver==1.4.8
LangSegment==0.3.5
lazy_loader==0.4
letta-client==0.1.167
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
modelscope==1.26.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.0
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.22.0
onnxruntime-gpu==1.22.0
openai==1.72.0
OpenCC==1.1.9
opencv-python==4.11.0.86
orjson==3.10.18
oss2==2.19.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
proces==0.1.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py3langid==0.3.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pynput==1.8.1
pyparsing==3.2.3
pypinyin==0.54.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-xlib==0.33
pytorch-lightning==2.5.1.post0
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.12
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
torch-complex==0.4.4
torchaudio==2.6.0
torchmetrics==1.7.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.2.0
typeguard==4.4.4
typer==0.16.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.14.0
tzdata==2025.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.0
vapi-server-sdk==1.5.1
Wave==0.0.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wordsegment==1.3.1
yarl==1.20.0
